fuzz: fix build (#48909)
* fuzz: fix build

A bug in the infra breaks if a fuzzer and benchmark is in one folder

See https://github.com/AdamKorcz/go-118-fuzz-build/pull/17

Split it out

* oops
